Move `repype.cli.run_cli` to `repype.__main__.run_cli`
xref #26
