+++ docs/source/repype.__main__.rst
@@ -1,0 +1,7 @@
+repype.__main__
+===============
+
+.. automodule:: repype.__main__
+    :members:
+    :undoc-members:
+    :show-inheritance:
